+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 - Voting system backend
 - Project collaboration backend
 - User profile and dashboard backend
+- Threaded comments backend
 
 ### Fixed
 
 - Getting posts by ID
+- Getting all posts endpoint
 
 ## [1.0.0] - 2026-05-23